Split `GitLike` trait from `Remote`
