+++ packages/ploys/src/repository/mod.rs
@@ -6,6 +6,7 @@
 mod cache;
 mod remote;
 mod spec;
+mod vcs;
 
 #[cfg(feature = "git")]
 pub mod git;
@@ -21,6 +22,7 @@ use std::path::Path;
 
 pub(crate) use self::remote::Remote;
 pub use self::spec::{Error as RepoSpecError, RepoSpec, ShortRepoSpec};
+pub use self::vcs::GitLike;
 
 /// Defines a file repository.
 pub trait Repository {
