+++ packages/ploys/src/repository/remote.rs
@@ -1,23 +1,15 @@
-use std::path::PathBuf;
-
 use semver::Version;
 
 use crate::changelog::Release;
 use crate::package::BumpOrVersion;
 
-use super::Repository;
+use super::GitLike;
 
 /// A remote repository.
 ///
 /// This defines the shared API of a remote repository to simplify feature flag
 /// handling.
-pub trait Remote: Repository {
-    /// Gets the commit SHA.
-    fn sha(&self) -> Result<String, Self::Error>;
-
-    /// Commits the changes to the repository.
-    fn commit(&self, message: &str, files: Vec<(PathBuf, String)>) -> Result<String, Self::Error>;
-
+pub trait Remote: GitLike {
     /// Requests a package release.
     fn request_package_release(
         &self,
@@ -32,15 +24,6 @@ pub trait Remote: Repository {
         version: &Version,
         is_primary: bool,
     ) -> Result<Release, Self::Error>;
-
-    /// Gets the default branch.
-    fn get_default_branch(&self) -> Result<String, Self::Error>;
-
-    /// Creates a new branch.
-    fn create_branch(&self, name: &str) -> Result<(), Self::Error>;
-
-    /// Updates the branch to point to the given SHA.
-    fn update_branch(&self, name: &str, sha: &str) -> Result<(), Self::Error>;
 
     /// Creates a pull request.
     fn create_pull_request(
@@ -67,14 +50,6 @@ impl<T> Remote for &T
 where
     T: Remote,
 {
-    fn sha(&self) -> Result<String, Self::Error> {
-        (*self).sha()
-    }
-
-    fn commit(&self, message: &str, files: Vec<(PathBuf, String)>) -> Result<String, Self::Error> {
-        (*self).commit(message, files)
-    }
-
     fn request_package_release(
         &self,
         package: &str,
@@ -90,18 +65,6 @@ where
         is_primary: bool,
     ) -> Result<Release, Self::Error> {
         (*self).get_changelog_release(package, version, is_primary)
-    }
-
-    fn get_default_branch(&self) -> Result<String, Self::Error> {
-        (*self).get_default_branch()
-    }
-
-    fn create_branch(&self, name: &str) -> Result<(), Self::Error> {
-        (*self).create_branch(name)
-    }
-
-    fn update_branch(&self, name: &str, sha: &str) -> Result<(), Self::Error> {
-        (*self).update_branch(name, sha)
     }
 
     fn create_pull_request(
